Added capabilities to set error codes and info arrays
Error codes and info arrays can now be set in a manner similar to setting
responses. The info array does not have to be set completely. Undefined
values will simply be returned as empty strings.

In addition, tests were added to show usage with multiple options and
different levels of specificity.