Fixed result selection bug
Previously, if you had (for instance) specified a response for 3 option
values and also specified an error code for 2 of those options, the error
code would not have been returned (because the algorithm would have
matched the 3 options but then would not have found a corresponding error
code). This commit includes a fix and a test to cover this case.

It also includes a refactoring to remove code duplication. That being said, the refactoring does not really make the code nicer.